+++ src/token.rs
@@ -195,7 +195,7 @@ impl Token {
         Annot::new(TokenKind::LineTerm, loc)
     }
 
-    pub fn new_eof(pos: usize) -> Self {
+    pub fn new_eof(pos: u32) -> Self {
         Annot::new(TokenKind::EOF, Loc(pos, pos))
     }
